com.utilities.extensions 1.1.6
- pulled out SerializedDictionaryObject class and made it public
- fixed height calculation for serialized dictionary property drawer
