+++ Utilities.Extensions/Packages/com.utilities.extensions/Editor/PropertyDrawers/SerializedDictionaryObject.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 64a99889d530a93469bc2100a2519663
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
